pull FL out of fixer, make configurable in a way that makes sense to me
