--- src/main/java/edu/lu/uni/serval/tbar/IFixer.java
@@ -1,22 +1,0 @@
-package edu.lu.uni.serval.tbar;
-
-import java.util.List;
-
-import edu.lu.uni.serval.tbar.AbstractFixer.SuspCodeNode;
-import edu.lu.uni.serval.tbar.utils.SuspiciousPosition;
-
-/**
- * Fixer Interface.
- * 
- * @author kui.liu
- *
- */
-public interface IFixer {
-
-	public List<SuspiciousPosition> readSuspiciousCodeFromFile();
-	
-	public List<SuspCodeNode> parseSuspiciousCode(SuspiciousPosition suspiciousCode);
-
-	public void fixProcess();
-	
-}
