+++ src/main/java/edu/lu/uni/serval/tbar/faultloc/PerfectFaultLoc.java
@@ -1,0 +1,50 @@
+package edu.lu.uni.serval.tbar.faultloc;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import edu.lu.uni.serval.tbar.config.Configuration;
+import edu.lu.uni.serval.tbar.utils.FileHelper;
+import edu.lu.uni.serval.tbar.utils.SuspiciousPosition;
+import edu.lu.uni.serval.tbar.dataprepare.DataPreparer;
+
+public class PerfectFaultLoc extends AbstractFaultLoc {
+    private List<SuspiciousPosition> suspiciousCodeList = new ArrayList<>();
+
+    public PerfectFaultLoc(DataPreparer d, String dataType, String buggyProject, String filePath) {
+        super(d,dataType,buggyProject);
+		String[] posArray = FileHelper.readFile(Configuration.knownBugPositions).split("\n");
+		Boolean isBuggyProject = null;
+		for (String pos : posArray) {
+			if (isBuggyProject == null || isBuggyProject) {
+				if (pos.startsWith(buggyProject + "@")) {
+					isBuggyProject = true;
+					
+					String[] elements = pos.split("@");
+	            	String[] lineStrArr = elements[2].split(",");
+	            	String classPath = elements[1];
+	            	String shortSrcPath = dp.srcPath.substring(dp.srcPath.indexOf(buggyProject) + buggyProject.length() + 1);
+	            	classPath = classPath.substring(shortSrcPath.length(), classPath.length() - 5);
+
+	            	for (String lineStr : lineStrArr) {
+	    				if (lineStr.contains("-")) {
+	    					String[] subPos = lineStr.split("-");
+	    					for (int line = Integer.valueOf(subPos[0]), endLine = Integer.valueOf(subPos[1]); line <= endLine; line ++) {
+	    						SuspiciousPosition sp = new SuspiciousPosition();
+	    		            	sp.classPath = classPath;
+	    		            	sp.lineNumber = line;
+	    		            	suspiciousCodeList.add(sp);
+	    					}
+	    				} else {
+	    					SuspiciousPosition sp = new SuspiciousPosition();
+	    	            	sp.classPath = classPath;
+	    	            	sp.lineNumber = Integer.valueOf(lineStr);
+	    	            	suspiciousCodeList.add(sp);
+	    				}
+	    			}
+				} else if (isBuggyProject!= null && isBuggyProject) isBuggyProject = false;
+			} else if (!isBuggyProject) break;
+		}
+    }
+
+}
